Move generalTable to components/generalTable folder
