+++ src/Routes/Repositories/RepositoryTable.js
@@ -1,83 +1,41 @@
-import React, { useState } from 'react';
-import ToolbarHeader from './ToolbarHeader';
-import ToolbarFooter from './ToolbarFooter';
-import Table from './Table';
+import React from 'react';
+import GeneralTable from '../../components/generalTable/GeneralTable';
 import PropTypes from 'prop-types';
-import EmptyState from './Empty';
 
-//const TableFilterContext = createContext({
-//  filters: [{ label: "Name", type: "text", key: "name" }],
-//});
+const filters = [
+  { label: 'Name', type: 'text' },
+  {
+    label: 'Distribution',
+    type: 'checkbox',
+    options: [{ option: '8.4' }, { option: '8.3' }],
+  },
+  {
+    label: 'Status',
+    type: 'checkbox',
+    options: [{ option: 'BUILDING' }, { option: 'CREATED' }],
+  },
+];
 
-const RepositoryTable = ({ data, toolbarButtons, toggle }) => {
-  const [input, setInput] = useState('');
-  const [sortDirection, setSortDirection] = useState('desc');
-  const [perPage, setPerPage] = useState(10);
-  const [page, setPage] = useState(1);
-
-  const filteredByName = () =>
-    data.filter((repo) =>
-      repo.Name.toLowerCase().includes(input.toLowerCase())
-    );
-
-  const sortedByDirection = (rows) =>
-    rows.sort((a, b) =>
-      sortDirection === 'asc'
-        ? a.name.toLowerCase().localeCompare(b.name.toLowerCase())
-        : b.name.toLowerCase().localeCompare(a.name.toLowerCase())
-    );
-
+const RepositoryTable = ({ data, openModal }) => {
   return (
-    <>
-      <ToolbarHeader
-        count={filteredByName().length}
-        toolbarButtons={toolbarButtons}
-        input={input}
-        setInput={setInput}
-        perPage={perPage}
-        setPerPage={setPerPage}
-        page={page}
-        setPage={setPage}
-      />
-      {filteredByName().length > 0 ? (
-        <Table
-          toggle={toggle}
-          columns={['Name']}
-          rows={sortedByDirection(filteredByName()).slice(
-            (page - 1) * perPage,
-            (page - 1) * perPage + perPage
-          )}
-          sortDirection={sortDirection}
-          setSortDirection={setSortDirection}
-        />
-      ) : (
-        <EmptyState
-          bgColor="white"
-          icon="search"
-          title="No match found"
-          secondaryActions={[
-            {
-              title: 'Clear all filters',
-              onClick: () => setInput(''),
-            },
-          ]}
-        />
-      )}
-      <ToolbarFooter
-        count={filteredByName().length}
-        setInput={setInput}
-        perPage={perPage}
-        setPerPage={setPerPage}
-        page={page}
-        setPage={setPage}
-      />
-    </>
+    <GeneralTable
+      filters={filters}
+      data={data}
+      actionFunction={openModal}
+      toolbarButtons={[
+        {
+          title: 'Add repository',
+          click: () => openModal({ type: 'add' }),
+        },
+      ]}
+    />
   );
 };
 RepositoryTable.propTypes = {
+  filters: PropTypes.func,
   data: PropTypes.array,
   toolbarButtons: PropTypes.array,
-  toggle: PropTypes.func,
+  openModal: PropTypes.func,
 };
 
 export default RepositoryTable;
